+++ ChangeLog.md
@@ -1,3 +1,7 @@
+- v1.0.1
+  + 修改 postinst ，为了安装目录权限
+  + 修改 debian 打包
+  
 - v1.0.0
   + 删除：
     - CLunarCalendar::AddAnniversary